info to csv
I have the information from the api put into a csv. I set the limit as 1000 but it seems to stop at 667 so that might be the limit per request of the website. example csv provided, fsym = BTC, tsym = USD, e=CCCAGG, toTs = 1417547600. I used the every hour api, but the difference between the first and final time is about 2 hours. So Im not exactly sure how the limit works. It might be how many results you want within that hour?