data for a year
every hour from 11/11/15 to 12/31/16 returned and put in a csv (time can be changed, and i didnt adjust it to fit perfectly in a year) This scraper should work for any of the history endpoints by changing the link. API also has a social media url so we can adapt this to that.