My ramblings about NN
tutorial followed: http://dacatay.com/data-science/part-6-time-series-prediction-neural-networks-python/
General observations the smaller the batch size the longer it takes (32 is pretty fast, 2 is significantly slower). Slight alterations to the layers has a major impact on loss. I also used some of the data I pulled, but ended up just using one attribute, since the tutorial only used one attrib. Not 100% sure how useful or practical this would be in predictions (or if I'm interpreting the results right). Added some images to show examples of the prediction (zoomin and zoomout is of the same graph with lag 6 and batch set to 2).