small additions
for some reason the column isolation doesn't seem to work so I altered the csv. I followed some examples, especially this one https://machinelearningmastery.com/display-deep-learning-model-training-history-in-keras/ , to try and get some more metrics, which I guess was a success. However, the added bits to get these metrics seem to affect the network as a whole. I don't know if this is farther away from a potential solution or not, just seems to be more parameters to fiddle with until a good combination is found.